requirements.txt
backports.weakref==1.0rc1
biopython==1.70
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
graphviz==0.8
h5py==2.7.1
html5lib==0.9999999
idna==2.6
ipython-genutils==0.2.0
jedi==0.11.0
jsonschema==2.6.0
jupyter-core==4.4.0
Keras==2.0.6
lxml==4.0.0
Markdown==2.6.8
nbformat==4.4.0
numpy==1.13.1
parso==0.1.0
plotly==2.2.1
protobuf==3.3.0
pydot==1.2.3
pyparsing==2.2.0
pyteomics==3.4.2
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.0
scipy==1.0.0
six==1.10.0
sklearn==0.0
tensorflow==1.2.1
Theano==0.9.0
traitlets==4.3.2
urllib3==1.22
Werkzeug==0.12.2
